US_ACQ removed. Improved realtime ACQ (if temperature)
08/02/2023
