+++ scan_dogbone_ACQ.py
@@ -40,7 +40,7 @@ PEref_path = os.path.join(MyDir, Experiment_PEref_file_name)
 WP_path = os.path.join(MyDir, Experiment_WP_file_name)
 Acqdata_path = os.path.join(MyDir, Experiment_acqdata_file_name)
 Temperature_path = os.path.join(MyDir, Experiment_Temperature_file_name)
-Experiment_description_path = os.path.join(MyDir, 'Experiment_description.txt')
+Experiment_description_path = os.path.join(MyDir, Experiment_description_file_name)
 Scanpath_path = os.path.join(MyDir, Experiment_scanpath_file_name)
 PEforCW_path = os.path.join(MyDir, Experiment_PEforCW_file_name)
 if not os.path.exists(MyDir):
@@ -53,10 +53,9 @@ print(f'Experiment path set to {MyDir}')
 ########################################################
 # Parameters and constants
 ########################################################
-# For Experiment_description do NOT use '\n'.
 # Suggestion: write material brand, model, dopong, etc. in Experiment_description
 Experiment_description = """Scanner test.
-Metacrylate dog-bone.
+Methacrylate dog-bone.
 Focused tx.
 Excitation_params: Pulse frequency (Hz).
 """
@@ -532,6 +531,9 @@ try:
             else:
                 means1[i] = tmp
         
+            Cwt = US.speedofsound_in_water(means1[i], method='Abdessamad', method_param=148)
+        
+        
         # -----------------------------------------------------------
         # Zero padding in case each channel has different scan length
         # -----------------------------------------------------------
@@ -543,16 +545,17 @@ try:
         
         # -----------------------------
         # Save temperature and acq data
-        # -----------------------------  
-        with open(Temperature_path, 'a') as f:
-            if twoSensors:
-                row = f'{means1[i]},{means2[i]},{Cw}'
-            else:
-                row = f'{means1[i]},{Cw}'
-            f.write(row+'\n')
+        # -----------------------------
+        if Temperature:
+            with open(Temperature_path, 'a') as f:
+                if twoSensors:
+                    row = f'{means1[i]},{means2[i]},{Cwt}'
+                else:
+                    row = f'{means1[i]},{Cwt}'
+                f.write(row+'\n')
         
-        _mode = 'wb' if i==0 else 'ab' # clear data from previous experiment before writing
         if Save_acq_data:
+            _mode = 'wb' if i==0 else 'ab' # clear data from previous experiment before writing
             with open(Acqdata_path, _mode) as f:
                 TT_Ascan.tofile(f)
                 PE_Ascan.tofile(f)
